\name{coexist-package}
\alias{coexist-package}
\alias{coexist}
\docType{package}
\title{
species coexistence modeling and analysis
}
\description{
species coexistence modeling and analysis under asymmetric dispersal and fluctuating source-sink dynamics;
testing the proportion of coexistence scenarios driven by neutral process and niche processes respectively.
}
\details{
\tabular{ll}{
Package: \tab coexist\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-08-01\cr
License: \tab GPL (>=2.0) \cr
LazyLoad: \tab yes\cr
}

}
\author{
Youhua Chen

Maintainer: 
Youhua Chen <yhchen@zoology.ubc.ca>
}
\references{
Chen YH (2012) coexist: an R package for performing species coexistence modeling and analysis under asymmetric dispersal and fluctuating source-sink dynamics.
http://code.google.com/p/coexist.
}
